requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
asyncpg==0.30.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt=3.2.2
python-dotenv==1.1.0
pydantic==2.11.4
alembic==1.15.2
opencv-python==4.9.0.80
tensorflow==2.16.1
keras==3.3.3